Update test class names
